+++ webapp/src/components/App/Singlestudy/explore/Configuration/AdequacyPatch/index.tsx
@@ -59,7 +59,6 @@ function AdequacyPatch() {
           ),
         },
       ]}
-      TabListProps={{ sx: { mt: -2 } }}
     />
   );
 }