+++ webapp/src/components/common/Form/index.tsx
@@ -45,6 +45,7 @@ import { SubmitHandlerPlus, UseFormReturnPlus } from "./types";
 import FormContext from "./FormContext";
 import useFormApiPlus from "./useFormApiPlus";
 import useFormUndoRedo from "./useFormUndoRedo";
+import { mergeSxProp } from "../../../utils/muiUtils";
 
 export interface AutoSubmitConfig {
   enable: boolean;
@@ -76,6 +77,7 @@ export interface FormProps<
   submitButtonIcon?: LoadingButtonProps["startIcon"];
   miniSubmitButton?: boolean;
   hideSubmitButton?: boolean;
+  hideFooterDivider?: boolean;
   onStateChange?: (state: FormState<TFieldValues>) => void;
   autoSubmit?: boolean | AutoSubmitConfig;
   enableUndoRedo?: boolean;
@@ -100,6 +102,7 @@ function Form<TFieldValues extends FieldValues, TContext>(
     submitButtonIcon = <SaveIcon />,
     miniSubmitButton,
     hideSubmitButton,
+    hideFooterDivider,
     onStateChange,
     autoSubmit,
     enableUndoRedo,
@@ -112,10 +115,12 @@ function Form<TFieldValues extends FieldValues, TContext>(
   const enqueueErrorSnackbar = useEnqueueErrorSnackbar();
   const { t } = useTranslation();
   const autoSubmitConfig = toAutoSubmitConfig(autoSubmit);
+
   const [showAutoSubmitLoader, setShowAutoSubmitLoader] = useDebouncedState(
     false,
     750,
   );
+
   const fieldAutoSubmitListeners = useRef<
     Record<string, ((v: any) => any | Promise<any>) | undefined>
   >({});
@@ -126,6 +131,7 @@ function Form<TFieldValues extends FieldValues, TContext>(
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   const submitSuccessfulCb = useRef(() => {});
   const preventClose = useRef(false);
+
   const contextValue = useMemo(
     () => ({ isAutoSubmitEnabled: autoSubmitConfig.enable }),
     [autoSubmitConfig.enable],
@@ -154,9 +160,9 @@ function Form<TFieldValues extends FieldValues, TContext>(
   // Don't add `isValid` because we need to trigger fields validation.
   // In case we have invalid default value for example.
   const isSubmitAllowed = isDirty && !isSubmitting;
-  const showSubmitButton = !hideSubmitButton && !autoSubmitConfig.enable;
-  const showFooter = showSubmitButton || enableUndoRedo;
   const rootError = errors.root?.[ROOT_ERROR_KEY];
+  const showSubmitButton = !hideSubmitButton && !autoSubmitConfig.enable;
+  const showFooter = showSubmitButton || enableUndoRedo || rootError;
 
   const formApiPlus = useFormApiPlus({
     formApi,
@@ -221,6 +227,9 @@ function Form<TFieldValues extends FieldValues, TContext>(
       if (preventClose.current) {
         // eslint-disable-next-line no-param-reassign
         event.returnValue = t("form.submit.inProgress");
+      } else if (isDirty) {
+        // eslint-disable-next-line no-param-reassign
+        event.returnValue = t("form.changeNotSaved");
       }
     };
 
@@ -229,13 +238,14 @@ function Form<TFieldValues extends FieldValues, TContext>(
     return () => {
       window.removeEventListener("beforeunload", listener);
     };
-  }, [t]);
+  }, [t, isDirty]);
 
   useUpdateEffect(() => onStateChange?.(formState), [formState]);
 
   useEffect(() => setRef(apiRef, formApiPlus));
 
   usePrompt(t("form.submit.inProgress"), preventClose.current);
+  usePrompt(t("form.changeNotSaved"), isDirty);
 
   ////////////////////////////////////////////////////////////////
   // Submit
@@ -323,13 +333,22 @@ function Form<TFieldValues extends FieldValues, TContext>(
   return (
     <Box
       {...formProps}
-      sx={sx}
+      sx={mergeSxProp(
+        {
+          display: "flex",
+          flexDirection: "column",
+          height: 1,
+          overflow: "auto",
+        },
+        sx,
+      )}
       component="form"
       onSubmit={handleFormSubmit}
       className={clsx("Form", className)}
     >
       {showAutoSubmitLoader && (
         <Box
+          className="Form__Loader"
           sx={{
             position: "sticky",
             top: 0,
@@ -337,65 +356,78 @@ function Form<TFieldValues extends FieldValues, TContext>(
             height: 0,
             textAlign: "right",
           }}
-          className="Form__Loader"
         >
           <CircularProgress color="secondary" size={20} sx={{ mr: 1 }} />
         </Box>
       )}
-      <FormContext.Provider value={contextValue}>
-        {RA.isFunction(children) ? (
-          children(formApiPlus)
-        ) : (
-          <FormProvider {...formApiPlus}>{children}</FormProvider>
-        )}
-      </FormContext.Provider>
-      {rootError && (
-        <Box color="error.main" sx={{ fontSize: "0.9rem", mb: 2 }}>
-          {rootError.message || t("form.submit.error")}
-        </Box>
-      )}
+      <Box className="Form__Content" sx={{ overflow: "auto" }}>
+        <FormContext.Provider value={contextValue}>
+          {RA.isFunction(children) ? (
+            children(formApiPlus)
+          ) : (
+            <FormProvider {...formApiPlus}>{children}</FormProvider>
+          )}
+        </FormContext.Provider>
+      </Box>
       {showFooter && (
-        <Box sx={{ display: "flex", pt: 1 }} className="Form__Footer">
-          {showSubmitButton && (
-            <>
-              <LoadingButton
-                type="submit"
-                disabled={!isSubmitAllowed}
-                loading={isSubmitting}
-                {...(miniSubmitButton
-                  ? {
-                      children: submitButtonIcon,
-                    }
-                  : {
-                      loadingPosition: "start",
-                      startIcon: submitButtonIcon,
-                      variant: "contained",
-                      children: submitButtonText || t("global.save"),
-                    })}
-              />
-              {enableUndoRedo && (
-                <Divider sx={{ mx: 2 }} orientation="vertical" flexItem />
-              )}
-            </>
+        <Box
+          className="Form__Footer"
+          sx={{
+            display: "flex",
+            flexDirection: "column",
+            gap: 1.5,
+            mt: hideFooterDivider ? 0 : 1.5,
+          }}
+        >
+          {!hideFooterDivider && <Divider flexItem />}
+          {rootError && (
+            <Box color="error.main" sx={{ fontSize: "0.9rem" }}>
+              {rootError.message || t("form.submit.error")}
+            </Box>
           )}
-          {enableUndoRedo && (
-            <>
-              <Tooltip title={t("global.undo")}>
-                <span>
-                  <IconButton onClick={undo} disabled={!canUndo}>
-                    <UndoIcon />
-                  </IconButton>
-                </span>
-              </Tooltip>
-              <Tooltip title={t("global.redo")}>
-                <span>
-                  <IconButton onClick={redo} disabled={!canRedo}>
-                    <RedoIcon />
-                  </IconButton>
-                </span>
-              </Tooltip>
-            </>
-          )}
+          <Box className="Form__Footer__Actions" sx={{ display: "flex" }}>
+            {showSubmitButton && (
+              <>
+                <LoadingButton
+                  type="submit"
+                  size="small"
+                  disabled={!isSubmitAllowed}
+                  loading={isSubmitting}
+                  {...(miniSubmitButton
+                    ? {
+                        children: submitButtonIcon,
+                      }
+                    : {
+                        loadingPosition: "start",
+                        startIcon: submitButtonIcon,
+                        variant: "contained",
+                        children: submitButtonText || t("global.save"),
+                      })}
+                />
+                {enableUndoRedo && (
+                  <Divider sx={{ mx: 2 }} orientation="vertical" flexItem />
+                )}
+              </>
+            )}
+            {enableUndoRedo && (
+              <>
+                <Tooltip title={t("global.undo")}>
+                  <span>
+                    <IconButton size="small" onClick={undo} disabled={!canUndo}>
+                      <UndoIcon />
+                    </IconButton>
+                  </span>
+                </Tooltip>
+                <Tooltip title={t("global.redo")}>
+                  <span>
+                    <IconButton size="small" onClick={redo} disabled={!canRedo}>
+                      <RedoIcon />
+                    </IconButton>
+                  </span>
+                </Tooltip>
+              </>
+            )}
+          </Box>
         </Box>
       )}
     </Box>